Hackathon_242VMTR01589/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
requests==2.31.0
PyQt5==5.15.9
pystray==0.19.4
Pillow==10.0.1
python-dotenv==1.0.0
pywin32==306
openai==1.1.1 